Change balance label to use doubles

+++ src/main/java/ledger/user_interface/ui_controllers/component/AccountBalanceLabel.java
@@ -80,7 +80,7 @@ public class AccountBalanceLabel extends Label implements IUIController, Initial
 
         String accountName = this.currentAccount.getName();
         if (accountName.length() > 20) accountName = accountName.substring(0, 17) + "...";
-        this.setText(accountName + ": " + String.format("%.2f", (float) net / 100.0));
+        this.setText(accountName + ": " + String.format("%.2f", (double) net / 100.0));
     }
 
     private void calculateBalanceForAllAccounts() {
@@ -107,6 +107,6 @@ public class AccountBalanceLabel extends Label implements IUIController, Initial
             net += transaction.getAmount();
         }
 
-        this.setText("All Accounts Balance: " + String.format("%.2f", (float) net / 100));
+        this.setText("All Accounts Balance: " + String.format("%.2f", (double) net / 100));
     }
 }
